requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-core
langchain-text-splitters
sentence-transformers
faiss-cpu
pypdf
python-dotenv
